fix(rust): lint the tests, and fix what that finds
CI ran `cargo clippy --all-features -- -D warnings`, which covers the library
and binaries and stops there. Test code had no lint floor at all, so a genuine
problem in a fixture builder had nowhere to surface.

Adding `--all-targets` reported 478 errors. 437 of them were `unwrap` and
`expect` in tests, which is what the crate's error-handling lints are for and
not what tests should be judged by: a test says what it expects and stops when
that is not what it got. Those three lints are allowed under `cfg(test)`, which
leaves every other lint watching test code — including the ones that were the
reason to look.

That left eleven, all real:

  - a `u16` cast to `u8` that truncates. 0o755 does not fit a byte; splitting it
    across permissions and permissions_high is the intent, so it now says so
    with to_le_bytes rather than relying on the cast to drop the high bits.
  - `MAGIC_TRAILER_SIZE as usize` where it is already a usize
  - a clone to build a one-element slice, where from_ref does not allocate
  - `.err().expect(...)` instead of `expect_err`
  - four `matches!(expr, true)` where the expression is the assertion
  - a `match` whose second arm is empty, which is an `if let`
  - a `format!` naming a variable it could interpolate
  - a function defined after the test module at the end of the file

The gate now sees them. Verified by putting a `len_zero` violation in a test
module and watching clippy report it:

    src/zz_probe.rs:10:12: error: length comparison to zero

`ci/quality-checks.sh` and the flavor-rs Makefile both pass `--all-targets`, so
`make lint` and CI agree about what is linted.

Closes #28.

+++ src/flavor-rs/src/api.rs
@@ -130,7 +130,7 @@ mod tests {
     use tempfile::tempdir;
 
     fn write_synthetic_pspf_package(path: &Path) {
-        let mut bytes = vec![0u8; MAGIC_TRAILER_SIZE as usize];
+        let mut bytes = vec![0u8; MAGIC_TRAILER_SIZE];
         bytes[..PACKAGE_EMOJI_BYTES.len()].copy_from_slice(PACKAGE_EMOJI_BYTES);
         let trailer_start = bytes.len() - MAGIC_WAND_EMOJI_BYTES.len();
         bytes[trailer_start..].copy_from_slice(MAGIC_WAND_EMOJI_BYTES);
